[sam] state-of-the-art commit

+++ hardware/sam/cores/sam/Print.cpp
@@ -43,10 +43,11 @@ void Print::write(const uint8_t *buffer, size_t size)
     write(*buffer++);
 }
 
-void Print::print(const String &s)
+void Print::print( const String &s )
 {
-  for (int i = 0; i < s.length(); i++) {
-    write(s[i]);
+  for ( int i = 0 ; i < (int)s.length() ; i++ )
+  {
+    write( s[i] ) ;
   }
 }
